feat(process.py): This commit implements the feature of having MessagePack as an option for output file
The process file is now able to output a MessagePack file (encoded from the JSON) to enable better efficiency when it's needed, update docs and requirements accordingly.

BREAKING CHANGE: To run a process, you should now provide the name and the extension of the audio file but for the output, only the name of the output file. You can also use the `--format` flag to output a msgpack file instead of JSON, as per your needs.